Added comments and small enhancements
Changed the normalization in ComputePosteriorgram so that each frame
sums to 1 as oppposed to the maximal element always being 1. Added a new
function to return a ranked list of indiced based on probability mass
instead to total number of times it is the highest ranked posterior in a
frame. Closes #20