Bug Fix and GeneratePronunciation Function
Discovered a bug in the Matrix transpose function. The previous function
only worked properly on Square matrices. The function is now likely
slower because of an extra data copy, but it is correct.

Began working on a function to Generate Pronunciations using the
PosteriogramGenerator and MultiBestPath code. The basic functionality is
now there, but it needs more work to make it more universal.