Optimized the ApproximateViterbiSet Code.
The approximate viterbi set now saves the backward probabilities. This
way we always know the score to finish any path. Now each step of the
process is really just a forced alignment and is significantly faster.

I also added an option to limit the size of the generated pronunciation.
Now we can force the pronunciations to be a reasonable length instead of
trying to hack something with the minimum number of frames setting.