Added code to handle clustering and decoding HMMs
Fixed the clustering code so that it will only cluster the triphones in
the HMM. GeneratePronunciations now uses the clustered HMMs to generate
the pronunciations. New code will soon make the MultiBestPath code
obselete.
Removed the gaussian sampling code. This is important because the
compiler no longer needs the std11 option.